refactor: moving files to new dirs to follow golang standard of project layout, hope everything work
