+++ .gitignore
@@ -89,3 +89,4 @@ secrets.yaml
 __debug_bin
 
 users.db
+/configs/data/users.db
