+++ cmd/web/static/css/globals.css
